# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus API",
    "version": "2.0.0",
    "description": "REST API for the Nexus learning platform — manage topics, reading, MCQs, notes, formulas, todos, and daily study plans."
  },
  "servers": [
    { "url": "/", "description": "Current deployment" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": { "type": "string", "minLength": 3 },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created, returns JWT token" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful login, returns JWT" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
    "/topics": {
      "get": {
        "tags": ["Topics"],
        "summary": "List all topics for the authenticated user",
        "responses": {
          "200": { "description": "Array of topics" }
        }
      },
      "post": {
        "tags": ["Topics"],
        "summary": "Create a new topic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "category_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Topic created" }
        }
      }
    },
    "/topics/{id}": {
      "get": {
        "tags": ["Topics"],
        "summary": "Get topic detail with formulas, code snippets, MCQs, notes, and TODO items",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Topic detail" },
          "404": { "description": "Topic not found" }
        }
      },
      "put": {
        "tags": ["Topics"],
        "summary": "Update topic title",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "title": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" }
        }
      },
      "delete": {
        "tags": ["Topics"],
        "summary": "Delete topic and associated data",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Deleted" }
        }
      }
    },
...
    "/reading/{id}": {
      "get": {
        "tags": ["Reading"],
        "summary": "Get a single reading block by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Reading block" },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/calendar/assign": {
      "post": {
        "tags": ["Calendar"],
        "summary": "Assign a topic to a date in the daily plan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["topic_id", "date"],
                "properties": {
                  "topic_id": { "type": "integer" },
                  "date": { "type": "string", "format": "date", "description": "YYYY-MM-DD, must not be in the past" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Assigned" },
          "400": { "description": "Validation error" }
        }
      }
    },
...
  }
}